LRCI-4706 Force floating point division

+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/FilePropagator.java
@@ -203,7 +203,7 @@ public class FilePropagator {
 			_busySlaves.size() + _errorSlaves.size() + _mirrorSlaves.size() +
 				_targetSlaves.size();
 
-		if ((_mirrorSlaves.size() / totalSlaveCount) < 0.5F) {
+		if ((_mirrorSlaves.size() / (float)totalSlaveCount) < 0.5F) {
 			throw new FilePropagatorRuntimeException(
 				this,
 				JenkinsResultsParserUtil.combine(
